added TorchSparseLogCategoricalDistribution entropy registration
